🧠 Academic Club (DSL) Projects/25-2_DSL_Modeling_NLP2_HospitalAgent/25-2_DSL_Modeling_NLP2_HospitalAgent/rag_doctor_agent/requirements.txt
openai>=1.35.0
pandas>=2.0.3
numpy>=1.23.0
pydantic>=2.6.1
pyyaml>=6.0.1
python-dotenv>=1.0.1
langgraph>=0.2.20
